rag_engine/requirements.txt
transformers
faiss-cpu
pdfplumber
torch==2.2.2
sentence-transformers
pymongo
python-dotenv
numpy<2
langchain
langchain-community
langchain-huggingface

#HTTP client libraries, if you need to make REST calls:
requests